LayoutTests/http/tests/security/contentSecurityPolicy/object-src-none-blocked-expected.txt

CONSOLE MESSAGE: Refused to load data:application/x-webkit-test-netscape,logifloaded because it does not appear in the object-src directive of the Content Security Policy.


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
